Motor Trend Car Road Tests

A version of the Motor Trend car road tests data from the datasets package, but as a tibble with categorical variables stored as factors.

Usage

mtcars2

Format

A tibble with 32 rows and 12 variables.

model

Make and model of the car

mpg

Miles/(US) gallon

cyl

Number of cylinders (as a factor)

disp

Displacement (cu.in.)

hp

Gross horsepower

drat

Rear axle ratio

wt

Weight (1000 lbs)

qsec

1/4 mile time

vs

Engine (V-shaped / straight)

am

Transmission (automatic / manual)

gear

Number of forward gears (as a factor)

carb

Number of carburetors (as a factor)

Source

The datasets package; see datasets::mtcars.